Cosmetic changes to keep Eclipse happy
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/branches/FOP_0-20-0_Alt-Design@197442 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/fop/fonts/FontUtil.java
@@ -35,7 +35,7 @@ public class FontUtil {
         int weight = 400;
         try {
             weight = Integer.parseInt(text);
-            weight = ((int)weight / 100) * 100;
+            weight = (weight / 100) * 100;
             weight = Math.max(weight, 100);
             weight = Math.min(weight, 900);
         } catch (NumberFormatException nfe) {
